spender: extract parent/node-view PSBT helpers from multifundchannel
Move update_parent_psbt/update_node_psbt out of openchannel.c into a
shared psbt_multiview module so the splice plugin can reuse them for
cross-channel splices (one funding tx backing several interactive-tx
negotiations at once).

Generalized signatures decouple them from the mfc structs:
- the funding-output special case (keep dest->funding_serial verbatim)
  becomes a caller-supplied preserve list, with a new symmetric list
  for inputs (a splice leg's old-funding input is daemon-minted with
  an even serial and must land on the parent unchanged; mfc passes
  NULL so its behavior is bit-identical);
- an optional parent_changed out-param reports whether the merge
  touched the parent (the splice flow uses it to decide which sibling
  negotiations need another update round).

The dead cleanup loops (`size_t i > -1` is always false, so they never
executed) are dropped with a comment explaining why freeing the copy
is safe without them.

Behavior of multifundchannel is unchanged; verified with
test_multifunding_v2_exclusive.
